Split up ERC20Tokens test into individual files.
