+++ tests/__init__.py
@@ -1,6 +1,7 @@
 import tests.test_Court
 import tests.test_Deploy
-import tests.test_ERC20Tokens
+import tests.test_ERC20Token
+import tests.test_ERC20FeeToken
 import tests.test_EtherNomin
 import tests.test_FeeCollection
 import tests.test_Havven
