+++ tests/test_ERC20FeeToken.py
@@ -12,155 +12,12 @@ ERC20FeeToken_SOURCE = "contracts/ERC20FeeToken.sol"
 
 
 def setUpModule():
-    print("Testing ERC20Tokens...")
+    print("Testing ERC20FeeToken...")
 
 
 def tearDownModule():
     print()
 
-
-class TestERC20Token(unittest.TestCase):
-    def setUp(self):
-        self.snapshot = take_snapshot()
-
-    def tearDown(self):
-        restore_snapshot(self.snapshot)
-
-    @classmethod
-    def setUpClass(cls):
-        cls.assertReverts = assertReverts
-
-        compiled = compile_contracts([ERC20Token_SOURCE])
-        cls.erc20_abi = compiled['ERC20Token']['abi']
-        cls.erc20_event_dict = generate_topic_event_map(cls.erc20_abi)
-        cls.erc20token, cls.construction_txr = attempt_deploy(compiled, 'ERC20Token', 
-                                                              MASTER, ["Test Token", "TEST", 
-                                                              1000 * UNIT, MASTER])
-
-        cls.totalSupply = lambda self: cls.erc20token.functions.totalSupply().call()
-        cls.name = lambda self: cls.erc20token.functions.name().call()
-        cls.symbol = lambda self: cls.erc20token.functions.symbol().call()
-        cls.balanceOf = lambda self, account: cls.erc20token.functions.balanceOf(account).call()
-        cls.allowance = lambda self, account, spender: cls.erc20token.functions.allowance(account, spender).call()
-
-        cls.transfer = lambda self, sender, to, value: mine_tx(cls.erc20token.functions.transfer(to, value).transact({'from' : sender}))
-        cls.approve = lambda self, sender, spender, value: mine_tx(cls.erc20token.functions.approve(spender, value).transact({'from' : sender}))
-        cls.transferFrom = lambda self, sender, fromAccount, to, value: mine_tx(cls.erc20token.functions.transferFrom(fromAccount, to, value).transact({'from' : sender}))
-
-    def test_constructor(self):
-        self.assertEqual(self.name(), "Test Token")
-        self.assertEqual(self.symbol(), "TEST")
-        self.assertEqual(self.totalSupply(), 1000 * UNIT)
-        self.assertEqual(self.balanceOf(MASTER), 1000 * UNIT)
-
-    def test_transfer(self):
-        sender = MASTER
-        sender_balance = self.balanceOf(sender)
-
-        receiver = fresh_account()
-        receiver_balance = self.balanceOf(receiver)
-        self.assertEqual(receiver_balance, 0)
-
-        value = 10 * UNIT
-        total_supply = self.totalSupply()
-
-        # This should fail because receiver has no tokens
-        self.assertReverts(self.transfer, receiver, sender, value)
-        tx_receipt = self.transfer(sender, receiver, value)
-        # Check event is emitted properly.
-        self.assertEqual(get_event_data_from_log(self.erc20_event_dict, tx_receipt.logs[0])['event'], "Transfer")
-        self.assertEqual(self.balanceOf(receiver), receiver_balance+value)
-        self.assertEqual(self.balanceOf(sender), sender_balance-value)
-
-        # transfers should leave the supply unchanged
-        self.assertEqual(self.totalSupply(), total_supply)
-
-        value = 1001 * UNIT
-        # This should fail because balance < value and balance > totalSupply
-        self.assertReverts(self.transfer, sender, receiver, value)
-
-        # 0 value transfers are allowed.
-        value = 0
-        pre_sender_balance = self.balanceOf(sender)
-        pre_receiver_balance = self.balanceOf(receiver)
-        tx_receipt = self.transfer(sender, receiver, value)
-        # Check event is emitted properly.
-        self.assertEqual(get_event_data_from_log(self.erc20_event_dict, tx_receipt.logs[0])['event'], "Transfer")
-        self.assertEqual(self.balanceOf(receiver), pre_receiver_balance)
-        self.assertEqual(self.balanceOf(sender), pre_sender_balance)
-
-        # It is also possible to send 0 value transfer from an account with 0 balance.
-        no_tokens = fresh_account()
-        self.assertEqual(self.balanceOf(no_tokens), 0)
-        tx_receipt = self.transfer(no_tokens, receiver, value)
-        # Check event is emitted properly.
-        self.assertEqual(get_event_data_from_log(self.erc20_event_dict, tx_receipt.logs[0])['event'], "Transfer")
-        self.assertEqual(self.balanceOf(no_tokens), 0)
-
-    def test_approve(self):
-        approver, spender = fresh_accounts(2)
-        approval_amount = 1 * UNIT
-
-        tx_receipt = self.approve(approver, spender, approval_amount)
-        # Check event is emitted properly.
-        self.assertEqual(get_event_data_from_log(self.erc20_event_dict, tx_receipt.logs[0])['event'], "Approval")
-        self.assertEqual(self.allowance(approver, spender), approval_amount)
-
-        # Any positive approval amount is valid, even greater than total_supply.
-        approval_amount = self.totalSupply() * 100
-        tx_receipt = self.approve(approver, spender, approval_amount)
-        # Check event is emitted properly.
-        self.assertEqual(get_event_data_from_log(self.erc20_event_dict, tx_receipt.logs[0])['event'], "Approval")
-        self.assertEqual(self.allowance(approver, spender), approval_amount)
-
-    def test_transferFrom(self):
-        approver = MASTER
-        spender, receiver = fresh_accounts(2)
-
-        approver_balance = self.balanceOf(approver)
-        spender_balance = self.balanceOf(spender)
-        receiver_balance = self.balanceOf(receiver) 
-
-        value = 10 * UNIT
-        total_supply = self.totalSupply()
-
-        # This fails because there has been no approval yet
-        self.assertReverts(self.transferFrom, spender, approver, receiver, value)
-
-        tx_receipt = self.approve(approver, spender, 2 * value)
-        # Check event is emitted properly.
-        self.assertEqual(get_event_data_from_log(self.erc20_event_dict, tx_receipt.logs[0])['event'], "Approval")
-        self.assertEqual(self.allowance(approver, spender), 2 * value)
-
-        self.assertReverts(self.transferFrom, spender, approver, receiver, 2 * value + 1)
-        tx_receipt = self.transferFrom(spender, approver, receiver, value)
-        # Check event is emitted properly.
-        self.assertEqual(get_event_data_from_log(self.erc20_event_dict, tx_receipt.logs[0])['event'], "Transfer")
-
-        self.assertEqual(self.balanceOf(approver), approver_balance - value)
-        self.assertEqual(self.balanceOf(spender), spender_balance)
-        self.assertEqual(self.balanceOf(receiver), receiver_balance + value)
-        self.assertEqual(self.allowance(approver, spender), value)
-        self.assertEqual(self.totalSupply(), total_supply)
-
-        # Empty the account
-        tx_receipt = self.transferFrom(spender, approver, receiver, value)
-        # Check event is emitted properly.
-        self.assertEqual(get_event_data_from_log(self.erc20_event_dict, tx_receipt.logs[0])['event'], "Transfer")
-
-        approver = fresh_account()
-        # This account has no tokens
-        approver_balance = self.balanceOf(approver) 
-        self.assertEqual(approver_balance, 0)
-        self.assertEqual(self.allowance(approver, spender), 0)
-
-        tx_receipt = self.approve(approver, spender, value)
-        # Check event is emitted properly.
-        self.assertEqual(get_event_data_from_log(self.erc20_event_dict, tx_receipt.logs[0])['event'], "Approval")
-        self.assertEqual(self.allowance(approver, spender), value)
-
-        # This should fail because the approver has no tokens.
-        self.assertReverts(self.transferFrom, spender, approver, receiver, value)
 
 
 class TestERC20FeeToken(unittest.TestCase):
